requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
unstructured
chromadb
openai
tiktoken

# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. 
# Leave this line commented out. 